refactor(app): move MainScreen out of features/ to app/
MainScreen is the BottomNav router shell that hosts the four top-level
tabs — it isn't a feature module of its own and doesn't fit the
data/domain/presentation Clean Architecture layout other features
follow. Relocating it under lib/app/ matches CLAUDE.md's rule that
app-wide structural pieces live there, and clears the only stray file
at the features/ root.
